+++ tests/unit/test_explanations_collection.py
@@ -120,17 +120,23 @@ class DummyExplanation:
     def reset(self):
         self.calls.append(("reset", None))
 
-    def super_explanations(self, **kwargs):
-        self.calls.append(("super", kwargs))
+    def super_explanations(self, only_ensured=False, include_potential=True, copy=True):
+        self.calls.append(("super", {"only_ensured": only_ensured, "include_potential": include_potential, "copy": copy}))
+        return self
 
-    def semi_explanations(self, **kwargs):
-        self.calls.append(("semi", kwargs))
+    def semi_explanations(self, only_ensured=False, include_potential=True, copy=True):
+        self.calls.append(("semi", {"only_ensured": only_ensured, "include_potential": include_potential, "copy": copy}))
+        return self
 
-    def counter_explanations(self, **kwargs):
-        self.calls.append(("counter", kwargs))
+    def counter_explanations(self, only_ensured=False, include_potential=True, copy=True):
+        self.calls.append(("counter", {"only_ensured": only_ensured, "include_potential": include_potential, "copy": copy}))
+        return self
 
-    def ensured_explanations(self):
-        self.calls.append(("ensured", None))
+    def ensured_explanations(self, include_potential=True, copy=True):
+        self.calls.append(("ensured", {"include_potential": include_potential, "copy": copy}))
+        return self
+        self.calls.append(("ensured", kwargs))
+        return self
 
     def rank_features(self, feature_weights, num_to_show=None):
         order = list(range(len(feature_weights)))